requirements.txt
lml==0.0.9
lxml==4.3.0
pdfkit==0.6.1
pyexcel==0.5.10
pyexcel-io==0.5.11
pyexcel-xlsxr==0.5.2
python-dateutil==2.8.0
six==1.12.0
texttable==1.6.0
